Fix conversion warning

+++ c/tools/brotli.c
@@ -847,11 +847,11 @@ static void PrintBytes(size_t value) {
   if (value < 1024) {
     fprintf(stderr, "%d B", (int)value);
   } else if (value < 1048576) {
-    fprintf(stderr, "%0.3f KiB", value / 1024.0);
+    fprintf(stderr, "%0.3f KiB", (double)value / 1024.0);
   } else if (value < 1073741824) {
-    fprintf(stderr, "%0.3f MiB", value / 1048576.0);
+    fprintf(stderr, "%0.3f MiB", (double)value / 1048576.0);
   } else {
-    fprintf(stderr, "%0.3f GiB", value / 1073741824.0);
+    fprintf(stderr, "%0.3f GiB", (double)value / 1073741824.0);
   }
 }
